+++ src/main/java/ru/top/project/model/Lecture.java
@@ -55,7 +55,7 @@ public class Lecture {
     public BigInteger getId() {
         return id;
     }
-    public void setId(int id) {
-        this.id = BigInteger.valueOf(id);
+    public void setId(BigInteger id) {
+        this.id = id;
     }
 }
